trash for infostore

+++ com.openexchange.file.storage.fs/src/com/openexchange/file/storage/fs/FSFileAccess.java
@@ -347,7 +347,15 @@ public class FSFileAccess implements FileStorageFileAccess, FileStorageEfficient
     }
 
     @Override
-    public List<IDTuple> removeDocument(List<IDTuple> ids, long sequenceNumber) throws OXException {
+    public List<IDTuple> removeDocument(final List<IDTuple> ids, final long sequenceNumber) throws OXException {
+        return removeDocument(ids, sequenceNumber, true);
+    }
+
+    @Override
+    public List<IDTuple> removeDocument(final List<IDTuple> ids, final long sequenceNumber, boolean hardDelete) throws OXException {
+        if (false == hardDelete) {
+            throw FileStorageExceptionCodes.NO_SUCH_FOLDER.create();
+        }
         for (IDTuple idTuple : ids) {
             toFile(idTuple.getFolder(), idTuple.getId()).delete();
         }
